+++ bilby/core/result.py
@@ -235,8 +235,9 @@ class Result(dict):
             elif k in self.parameter_labels:
                 latex_labels.append(k)
             else:
-                raise ValueError('key {} not a parameter label or latex label'
-                                 .format(k))
+                logger.info(
+                    'key {} not a parameter label or latex label'.format(k))
+                latex_labels.append(' '.join(k.split('_')))
         return latex_labels
 
     @property
@@ -588,39 +589,6 @@ class Result(dict):
                 else:
                     self.prior_values[key]\
                         = priors[key].prob(self.posterior[key].values)
-
-    def construct_cbc_derived_parameters(self):
-        """ Construct widely used derived parameters of CBCs """
-        self.posterior['mass_chirp'] = (
-            (self.posterior.mass_1 * self.posterior.mass_2) ** 0.6 / (
-                self.posterior.mass_1 + self.posterior.mass_2) ** 0.2)
-        self.search_parameter_keys.append('mass_chirp')
-        self.parameter_labels.append('$\mathcal{M}$')
-
-        self.posterior['q'] = self.posterior.mass_2 / self.posterior.mass_1
-        self.search_parameter_keys.append('q')
-        self.parameter_labels.append('$q$')
-
-        self.posterior['eta'] = (
-            (self.posterior.mass_1 * self.posterior.mass_2) / (
-                self.posterior.mass_1 + self.posterior.mass_2) ** 2)
-        self.search_parameter_keys.append('eta')
-        self.parameter_labels.append('$\eta$')
-
-        self.posterior['chi_eff'] = (
-            (self.posterior.a_1 * np.cos(self.posterior.tilt_1) +
-                self.posterior.q * self.posterior.a_2 *
-                np.cos(self.posterior.tilt_2)) / (1 + self.posterior.q))
-        self.search_parameter_keys.append('chi_eff')
-        self.parameter_labels.append('$\chi_{\mathrm eff}$')
-
-        self.posterior['chi_p'] = (
-            np.maximum(self.posterior.a_1 * np.sin(self.posterior.tilt_1),
-                       (4 * self.posterior.q + 3) / (3 * self.posterior.q + 4) *
-                       self.posterior.q * self.posterior.a_2 *
-                       np.sin(self.posterior.tilt_2)))
-        self.search_parameter_keys.append('chi_p')
-        self.parameter_labels.append('$\chi_{\mathrm p}$')
 
     def check_attribute_match_to_other_object(self, name, other_object):
         """ Check attribute name exists in other_object and is the same