+++ examples/injection_examples/change_sampled_parameters.py
@@ -29,22 +29,25 @@ waveform_arguments = dict(waveform_approximant='IMRPhenomPv2',
                           reference_frequency=50.)
 
 # Create the waveform_generator using a LAL BinaryBlackHole source function
+# We specify a function which transforms a dictionary of parameters into the
+# appropriate parameters for the source model.
 waveform_generator = bilby.gw.waveform_generator.WaveformGenerator(
     sampling_frequency=sampling_frequency, duration=duration,
     frequency_domain_source_model=bilby.gw.source.lal_binary_black_hole,
-    parameter_conversion=
-        bilby.gw.conversion.convert_to_lal_binary_black_hole_parameters,
+    parameter_conversion=bilby.gw.conversion.convert_to_lal_binary_black_hole_parameters,
     waveform_arguments=waveform_arguments)
 
 # Set up interferometers.
 ifos = bilby.gw.detector.InterferometerList(['H1', 'L1', 'V1', 'K1'])
 ifos.set_strain_data_from_power_spectral_densities(
     sampling_frequency=sampling_frequency, duration=duration,
-    start_time=injection_parameters['geocent_time']-3)
+    start_time=injection_parameters['geocent_time'] - 3)
 ifos.inject_signal(waveform_generator=waveform_generator,
                    parameters=injection_parameters)
 
 # Set up prior
+# Note it is possible to sample in different parameters to those that were
+# injected.
 priors = bilby.gw.prior.BBHPriorSet()
 priors.pop('mass_1')
 priors.pop('mass_2')
@@ -69,9 +72,10 @@ likelihood = bilby.gw.likelihood.GravitationalWaveTransient(
     interferometers=ifos, waveform_generator=waveform_generator)
 
 # Run sampler
+# Note we've added a post-processing conversion function, this will generate
+# many useful additional parameters, e.g., source-frame masses.
 result = bilby.core.sampler.run_sampler(
     likelihood=likelihood, priors=priors, sampler='dynesty', outdir=outdir,
     injection_parameters=injection_parameters, label='DifferentParameters',
     conversion_function=bilby.gw.conversion.generate_all_bbh_parameters)
 result.plot_corner()
-