+++ examples/injection_examples/create_your_own_time_domain_source_model.py
@@ -1,41 +1,41 @@
 #!/usr/bin/env python
 """
 A script to show how to create your own time domain source model.
-A simple damped Gaussian signal is defined in the time domain, injected into noise in
-two interferometers (LIGO Livingston and Hanford at design sensitivity),
-and then recovered.
+A simple damped Gaussian signal is defined in the time domain, injected into
+noise in two interferometers (LIGO Livingston and Hanford at design
+sensitivity), and then recovered.
 """
 
-import bilby
 import numpy as np
-
+import bilby
 
 
 # define the time-domain model
-def time_domain_damped_sinusoid(time, amplitude, damping_time, frequency, phase, ra, dec, psi, geocent_time):
+def time_domain_damped_sinusoid(
+        time, amplitude, damping_time, frequency, phase):
     """
-    This example only creates a linearly polarised signal with only plus polarisation.
+    This example only creates a linearly polarised signal with only plus
+    polarisation.
     """
-
-    plus = amplitude * np.exp(-time / damping_time) * np.sin(2.*np.pi*frequency*time + phase)
+    plus = amplitude * np.exp(-time / damping_time) *\
+        np.sin(2 * np.pi * frequency * time + phase)
     cross = np.zeros(len(time))
-
     return {'plus': plus, 'cross': cross}
+
 
 # define parameters to inject.
 injection_parameters = dict(amplitude=5e-22, damping_time=0.1, frequency=50,
-                            phase=0,
-                            ra=0, dec=0, psi=0, geocent_time=0.)
+                            phase=0, ra=0, dec=0, psi=0, geocent_time=0.)
 
 duration = 0.5
 sampling_frequency = 2048
-outdir='outdir'
-label='time_domain_source_model'
+outdir = 'outdir'
+label = 'time_domain_source_model'
 
 # call the waveform_generator to create our waveform model.
-waveform = bilby.gw.waveform_generator.WaveformGenerator(duration=duration, sampling_frequency=sampling_frequency,
-                                                         time_domain_source_model=time_domain_damped_sinusoid,
-                                                         parameters=injection_parameters)
+waveform = bilby.gw.waveform_generator.WaveformGenerator(
+    duration=duration, sampling_frequency=sampling_frequency,
+    time_domain_source_model=time_domain_damped_sinusoid)
 
 # inject the signal into three interferometers
 ifos = bilby.gw.detector.InterferometerList(['H1', 'L1'])
@@ -47,7 +47,7 @@ ifos.inject_signal(waveform_generator=waveform,
 
 #  create the priors
 prior = injection_parameters.copy()
-prior['amplitude'] = bilby.core.prior.Uniform(1e-23, 1e-21, r'$h_0$')
+prior['amplitude'] = bilby.core.prior.LogUniform(1e-23, 1e-21, r'$h_0$')
 prior['damping_time'] = bilby.core.prior.Uniform(
     0, 1, r'damping time', unit='$s$')
 prior['frequency'] = bilby.core.prior.Uniform(0, 200, r'frequency', unit='Hz')
@@ -57,9 +57,8 @@ prior['phase'] = bilby.core.prior.Uniform(-np.pi / 2, np.pi / 2, r'$\phi$')
 likelihood = bilby.gw.likelihood.GravitationalWaveTransient(ifos, waveform)
 
 # launch sampler
-result = bilby.core.sampler.run_sampler(likelihood, prior, sampler='dynesty', npoints=1000,
-                                        injection_parameters=injection_parameters,
-                                        outdir=outdir, label=label)
+result = bilby.core.sampler.run_sampler(
+    likelihood, prior, sampler='dynesty', npoints=1000,
+    injection_parameters=injection_parameters, outdir=outdir, label=label)
 
 result.plot_corner()
-