+++ examples/other_examples/occam_factor_example.py
@@ -84,8 +84,8 @@ class Polynomial(bilby.Likelihood):
 
     def log_likelihood(self):
         res = self.y - self.polynomial(self.x, self.parameters)
-        return -0.5 * (np.sum((res / self.sigma)**2)
-                       + self.N*np.log(2*np.pi*self.sigma**2))
+        return -0.5 * (np.sum((res / self.sigma)**2) +
+                       self.N * np.log(2 * np.pi * self.sigma**2))
 
 
 def fit(n):